Require installation verification
This commit prevents a user from hitting the application without entering
a verification token.

Without this commit, the following is possible:

1. A user who has not enabled Authy yet hits the application
2. The user enters a phone number to turn Authy on
3. Instead of entering a token at the `verify-installation` endpoint, the
user navigates to another page.

This was only an issue for that first session where Authy is first enabled. After
entering a phone number and enabling Authy, subsequent sessions require an Authy
token.