Add support for phone call verification
It appears that getting a token via phone call isn't a standard part
of Authy, but is something they offer and support. This commit adds
a view helper, route, and controller action so that all of the pieces
to support phone call tokens are there. With this in the gem,
developers who want to offer phone call verification can add it easily
by extending the partials supplied by this gem.