Removes Jeweler.
This means we need to administer the gemspec directly, but in modern gemspecs that is more straightforward.

The new gemspec was built based on the template that bundler creates when you generate a new gem. All the dependencies were brought from the Gemfile into the gemspec and the other spec attributes transformed from the Jeweler config in the Rakefile.