Stops making API requests from tests.
It appears that this test suite was making requests to the sandbox API URL. This was recently causing time outs. This update adds webmock to block network requests from tests and uses rspec mocks to stub the response of the Authy library and test this implementation in a more robust and network independent manner.