Stop calling signed_in? before needed.
check_request_and_redirect_to_verify_token is guarded by is_signing_in? but both of them make calls to signed_in? which in turn calls warden.authenticate? which actually signs a user in if the right details are in the request. check_request_and_redirect_to_verify_token actually calls signed_in? and means to, so it is not necessary to call signed_in?, particularly so early in the guard which is run on every controller action, in is_signing_in?.

Fixes #76.